Sec: fix ineffective overflow guard for multisell quantities

+++ java/org/l2jmobius/gameserver/network/clientpackets/MultiSellChoose.java
@@ -169,7 +169,7 @@ public class MultiSellChoose extends ClientPacket
 						// this happens if 1 list entry has the same ingredient twice (example 2 swords = 1 dual)
 						if ((ex.getItemId() == e.getItemId()) && (ex.getEnchantLevel() == e.getEnchantLevel()))
 						{
-							if ((ex.getItemCount() + e.getItemCount()) > Integer.MAX_VALUE)
+							if (((long) ex.getItemCount() + e.getItemCount()) > Integer.MAX_VALUE)
 							{
 								player.sendPacket(SystemMessageId.YOU_HAVE_EXCEEDED_THE_QUANTITY_THAT_CAN_BE_INPUTTED);
 								return;
@@ -194,7 +194,7 @@ public class MultiSellChoose extends ClientPacket
 				// now check if the player has sufficient items in the inventory to cover the ingredients' expences
 				for (Ingredient e : ingredientsList)
 				{
-					if ((e.getItemCount() * _amount) > Integer.MAX_VALUE)
+					if (((long) e.getItemCount() * _amount) > Integer.MAX_VALUE)
 					{
 						player.sendPacket(SystemMessageId.YOU_HAVE_EXCEEDED_THE_QUANTITY_THAT_CAN_BE_INPUTTED);
 						return;
